requirements.txt
# OpenClaw — shared requirements.
# Broker: fastapi, uvicorn, pydantic
# Runner: requests
# Bot: discord.py, requests
# Tests: pytest

fastapi
uvicorn[standard]
pydantic
requests
discord.py
pytest
